requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.26

# Research and search tools
arxiv>=1.4.0
wikipedia>=1.4.0
tavily-python>=0.3.0
duckduckgo-search>=3.9.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Optional: For enhanced features
requests>=2.31.0
beautifulsoup4>=4.12.0